JNA callback should check inputs and outputs for null.
This addresses https://github.com/extism/java-sdk/issues/27
once we merge https://github.com/extism/extism/pull/760
and release a new version of libextism.

This PR simply check the parameters for null and the counts
to be valid, in preparation for https://github.com/extism/extism/pull/760
which otherwise would cause a NullPointerException when
outputs and inputs are empty.

Signed-off-by: Edoardo Vacchi <evacchi@users.noreply.github.com>